+++ backend/src/main/java/gov/cdc/usds/simplereport/api/heathcheck/BackendAndDatabaseHealthIndicator.java
@@ -1,8 +1,6 @@
 package gov.cdc.usds.simplereport.api.heathcheck;
 
-import com.okta.sdk.resource.client.ApiException;
 import gov.cdc.usds.simplereport.db.repository.FeatureFlagRepository;
-import gov.cdc.usds.simplereport.idp.repository.OktaRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.exception.JDBCConnectionException;
@@ -15,27 +13,15 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class BackendAndDatabaseHealthIndicator implements HealthIndicator {
   private final FeatureFlagRepository _ffRepo;
-  private final OktaRepository _oktaRepo;
-  public static final String ACTIVE_LITERAL = "ACTIVE";
 
   @Override
   public Health health() {
     try {
       _ffRepo.findAll();
-      String oktaStatus = _oktaRepo.getApplicationStatusForHealthCheck();
-
-      if (!ACTIVE_LITERAL.equals(oktaStatus)) {
-        log.info("Okta status didn't return ACTIVE, instead returned " + oktaStatus);
-        return Health.down().build();
-      }
       return Health.up().build();
 
       // reach into the ff repository returned a bad value or db connection issue respectively
     } catch (IllegalArgumentException | JDBCConnectionException e) {
-      return Health.down().build();
-    } catch (ApiException e) {
-      // Okta API call errored
-      log.info(e.getMessage());
       return Health.down().build();
     }
   }
